backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
groq==0.11.0
pdfplumber==0.11.0
python-docx==1.1.0
reportlab==4.0.7
gunicorn==21.2.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
